+++ application/dash_application/views/single_lane_ts.py
@@ -3,6 +3,7 @@ from collections import defaultdict
 import dash_html_components as html
 from dash.dependencies import Input, Output, State
 from ..dash_id import init_ids
+import dash_bootstrap_components as dbc
 from ..utility.plot_builder import *
 from ..utility.table_builder import table_tabs_single_lane, cutoff_table_data_ius
 from ..utility import df_manipulation as util
@@ -25,6 +26,9 @@ ids = init_ids([
     'approve-run-button',
     'miso-request-body',
     'miso-button',
+
+    # Alerts
+    "alerts-unknown-run",
 
     # Sidebar controls
     'all-runs',
@@ -291,11 +295,13 @@ def dataversion():
 
 def layout(query_string):
     query = sidebar_utils.parse_query(query_string)
+    unknown_runs = []  # Runs from the query string that do not exist
     # initial runs: should be empty unless query requests otherwise:
     #  * if query.req_run: use query.req_run
     #  * if query.req_start/req_end: use all runs, so that the start/end filters will be applied
     if "req_runs" in query and query["req_runs"]:
         initial["runs"] = query["req_runs"]
+        unknown_runs = [x for x in initial["runs"] if x not in ALL_RUNS]
     elif "req_start" in query and query["req_start"]:
         initial["runs"] = ALL_RUNS
         query["req_runs"] = ALL_RUNS  # fill in the runs dropdown
@@ -312,6 +318,13 @@ def layout(query_string):
     return core.Loading(fullscreen=True, type="dot", children=[
         html.Div(className='body', children=[
             html.Div(className="row jira-buttons", children=[
+                dbc.Alert(
+                    "Requested run does not exist: {}".format(unknown_runs),
+                    ids['alerts-unknown-run'],
+                    color="danger",
+                    dismissable=True,
+                    is_open=len(unknown_runs) > 0
+                ),
                 sidebar_utils.jira_button("Open an issue",
                                           ids['general-jira-issue-button'],
                                           {"display": "inline-block"},